BMP temperature field was too small
uin8_t vs uint16_t

+++ common/devices/include/BMP180.h
@@ -11,7 +11,7 @@ struct BMP180Config {
 
 struct BMP180Data {
 	uint32_t pressure;
-	uint8_t temperature;
+	uint16_t temperature;
 };
 
 extern EventSource BMP180DataEvt;
